requirements.txt
aiohttp==3.7.4
python-telegram-bot==13.7
requests==2.31.0

